Updates image population logic on card#show
In the case where a card has a large number of images attached, creating
image elements for all of them, and populating the DOM all at once will
crash in iOS 5.1.1 Safari.
With this change, an upper limit is set to
22500 pixels, where images are only added if the document doesn't grow
above this size.
This unfortunately also means, that not all images will be show each
time, but since the images are shuffled, some that aren't shown in one
instance, will probably be shown another time around.