Updates state transition for imagesLoaded
This changes the logic for when to start fading out the cover and show
the images. Instead of waiting for all images to load, trigger the state
change already after five images have been loaded. This should be enough
to cover the screen, whilst the remaining out-of-screen images can
continue loading in the background.