Adds MimeType gaurd to Card GetImages
Since only images will have previews generated for them, in case a file
of another type is attached, to the card it would trigger an 'slice
bounds out of range' error in GetPreviews.

With this change, only attachments of an image/ mimetype is included
with GetImages.

Even though the codepath should never be visited due to the above
change, GetPreviews is now also guarded, in case it's called where
there's no previews on an image.